backend/requirements.txt
fastapi
uvicorn[standard]
google-cloud-storage
google-cloud-compute
google-cloud-resource-manager
pg8000
sqlalchemy
python-dotenv
